+++ packages/api/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.18-unstable.2"></a>
+## [1.0.18-unstable.2](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.18-unstable.1...@aws-amplify/api@1.0.18-unstable.2) (2018-10-08)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.18-unstable.1"></a>
+## [1.0.18-unstable.1](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.18-unstable.0...@aws-amplify/api@1.0.18-unstable.1) (2018-10-05)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.18-unstable.0"></a>
+## [1.0.18-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.17-unstable.1...@aws-amplify/api@1.0.18-unstable.0) (2018-10-05)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.17"></a>
+## [1.0.17](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.17-unstable.1...@aws-amplify/api@1.0.17) (2018-10-04)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.17-unstable.1"></a>
+## [1.0.17-unstable.1](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.17-unstable.0...@aws-amplify/api@1.0.17-unstable.1) (2018-10-03)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.17-unstable.0"></a>
+## [1.0.17-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.16-unstable.4...@aws-amplify/api@1.0.17-unstable.0) (2018-10-03)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.16"></a>
+## [1.0.16](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.16-unstable.4...@aws-amplify/api@1.0.16) (2018-10-03)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.16-unstable.4"></a>
+## [1.0.16-unstable.4](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.16-unstable.3...@aws-amplify/api@1.0.16-unstable.4) (2018-10-01)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
 <a name="1.0.16-unstable.3"></a>
 ## [1.0.16-unstable.3](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.16-unstable.2...@aws-amplify/api@1.0.16-unstable.3) (2018-09-28)